requirements.txt
python-dotenv
Flask
Flask-Login
langchain
langchain-community
langchain-groq
nltk
SpeechRecognition
pydub
